Rename error.ts to showToastOnError.ts
